+++ Tile3D/File/TFileDir.cpp
@@ -464,7 +464,7 @@ bool TFileDir::NormalizeFileName(char* szFileName)
 	return true;
 }
 
-static const int CRC32Table[] =
+static const unsigned long CRC32Table[] =
 {
 	0x00000000L, 0x77073096L, 0xee0e612cL, 0x990951baL, 0x076dc419L,
 	0x706af48fL, 0xe963a535L, 0x9e6495a3L, 0x0edb8832L, 0x79dcb8a4L,
